refactor (refacotring in components)
